Move non-scripts into separate components folder
