+++ examples/typescript/financial_report/components/access_control/adminIdentity.ts
