+++ examples/typescript/financial_report/components/access_control/advisorIdentity.ts
